+++ messages/de/base.php
@@ -63,4 +63,5 @@ return [
     'Display monochrome toolbar header' => 'Monochromen Kopfbereich der Symbolleiste anzeigen',
     'Error when trying to connect ({error})' => 'Fehler beim Anschließen ({error})',
     'Mixed Active Content is not allowed. HTTPS address for ONLYOFFICE Docs is required.' => 'Mixed Active Content ist nicht möglich. HTTPS-Adresse für ONLYOFFICE Docs ist erforderlich.',
+    'Keep intermediate versions when editing (forcesave)' => 'Zwischenversionen bei der Bearbeitung aufbewahren (force save)',
 ];