+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Added
 - editor customization
 - validation of server settings on the settings page
+- keep intermediate versions when editing (forcesave)
+- connect to the demo server
+- trial period of 30 days for the demo server
+- transfer user region in conversion
 
 ## 2.4.0
 ## Added